Update highlight style to look like a drop shadow
The dashed border color actually changes between mainland and island
with island using black by default, which we thought was an unused color
so we used it before. On the island you could no longer tell which
highlight was for default stack and which was for reverse stack.

We experimented with the inverse color but white on the island is
almost invisible due to the lighter background. Instead we now opt
for a solid border and make it gray ish in every world so it looks like
a drop shadow