Remove auto stack to bottom
This change seemed to make sense but ultimately it is decided to not
include in this mod due to it being significantly more complex and
broad (operating on StackSend / Bounce, impacting many gameplay
situations) than just enabling it on manual drag.

In addition to the complexity it also makes harvesting apple trees
and berry bushes possible without any manual interaction from the player
apart from planting the apples/berries. Since the new apple trees and
berry bushes will now be auto stacked to the bottom of an existing
stack you do not have to do anything anymore. That is not how it should
be.