Fixed stacking to bottom of some cards that should not be allowed
For example, it was possible to stack Magic Glue to the bottom of
Baby Dragon but that should not be possible because you cannot
create that stack manually by dragging Baby Dragon on top of Magic
Glue --  because Baby Dragon cannot be dragged by the player. We
added a general check if the target card can be dragged itself so
this solves the described case as well as some potential others.